pyproject.toml
[project]
name = "rh_hume_project"
version = "0.1.0"
description = "Contested Logistics Intelligence Platform - The Hume Project"
readme = "README.md"
license = {text = "Proprietary"}
requires-python = ">=3.11"
authors = [
    {name = "Craig", email = "craig@example.com"}
]
keywords = ["logistics", "intelligence", "neo4j", "kafka", "hume", "ai"]

dependencies = [
    # Core
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    
    # API
    "fastapi>=0.109",
    "uvicorn[standard]>=0.27",
    "httpx>=0.26",
    "websockets>=12.0",
    
    # Data Processing
    "pandas>=2.1",
    "deltalake>=0.15",
    "pyarrow>=14.0",
    "numpy>=1.26",
    
    # Neo4j
    "neo4j>=5.15",
    
    # Kafka
    "kafka-python>=2.0",
    
    # LLM / AI
    "anthropic>=0.18",
    "openai>=1.10",
    "tiktoken>=0.5",
    
    # Embeddings (local option)
    "sentence-transformers>=2.3",
    
    # Utilities
    "tenacity>=8.2",
    "structlog>=24.1",
    "python-json-logger>=2.0",
    "rich>=13.7",
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.2",
    "mypy>=1.8",
    "pre-commit>=3.6",
    "ipython>=8.20",
    "jupyter>=1.0",
]

databricks = [
    "pyspark>=3.5",
    "databricks-sdk>=0.17",
]

[project.scripts]
hume-cli = "jobs.runner:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "jobs"]

# =============================================================================
# Ruff Configuration
# =============================================================================
[tool.ruff]
target-version = "py312"
line-length = 100
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
    ".eggs",
    "*.egg-info",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # pyflakes
    "I",      # isort
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
    "RUF",    # ruff-specific
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in default argument
    "SIM108", # ternary operator
]

[tool.ruff.lint.isort]
known-first-party = ["src", "jobs"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

# =============================================================================
# MyPy Configuration
# =============================================================================
[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_ignores = true
warn_redundant_casts = true
disallow_untyped_defs = false  # Gradual typing
ignore_missing_imports = true
exclude = [
    "tests/",
    "notebooks/",
]

[[tool.mypy.overrides]]
module = "kafka.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "neo4j.*"
ignore_missing_imports = true

# =============================================================================
# Pytest Configuration
# =============================================================================
[tool.pytest.ini_options]
minversion = "8.0"
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "unit: Unit tests (fast, no external dependencies)",
    "integration: Integration tests (requires services)",
    "golden: Golden test cases for agent quality",
    "slow: Slow tests (skip with -m 'not slow')",
]

# =============================================================================
# Coverage Configuration
# =============================================================================
[tool.coverage.run]
source = ["src", "jobs"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
